% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extract_samples}
\alias{extract_samples}
\title{Extract MCMC Samples with Burn-in and Thinning}
\usage{
extract_samples(x, parameter = "log_lik", burnin = 0, thin = 1)
}
\arguments{
\item{x}{An object of class "mrdltm_mcmc".}

\item{parameter}{Name of the parameter group.}

\item{burnin}{Number of iterations to discard. Defaults to 0.}

\item{thin}{Interval for thinning. Defaults to 1.}
}
\value{
A 3D array (iteration, chain, parameter).
}
\description{
Extract MCMC Samples with Burn-in and Thinning
}
