% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_toy_data.R
\name{generate_toy_data}
\alias{generate_toy_data}
\title{Generate Toy Data for MR-DLTM}
\usage{
generate_toy_data(
  n_cust = 10,
  n_item = 50,
  n_topic = 3,
  length_time = 30,
  n_var = 2,
  p_dim = 1,
  purchase_prob = 0.05
)
}
\arguments{
\item{n_cust}{Number of customers (C)}

\item{n_item}{Number of items (I)}

\item{n_topic}{Number of latent topics (Z)}

\item{length_time}{Length of time points (T)}

\item{n_var}{Number of marketing covariates including intercept (M)}

\item{p_dim}{Dimension of customer covariates including intercept (P)}

\item{purchase_prob}{Degree of sparcity}
}
\value{
A list containing:
\itemize{
\item observations: List of observed data (data, x_it, Dc)
\item true_params: List of ground truth parameters
}
}
\description{
Generates synthetic data based on the Market Response Dynamic Linear Topic Model(MR-DLTM).
}
