% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrdltm_mcmc.R
\name{mrdltm_mcmc}
\alias{mrdltm_mcmc}
\title{Gibbs Sampler for MR-DLTM}
\usage{
mrdltm_mcmc(model, iter = 2000, burnin = 1000, store_z = FALSE, quiet = TRUE)
}
\arguments{
\item{model}{An object of class "mrdltm_model".}

\item{iter}{Total MCMC iterations.}

\item{burnin}{Burn-in iterations.}

\item{store_z}{A flag to store z_cit}

\item{quiet}{A flag to show a progress-bar}
}
\value{
A list of class "mrdltm_mcmc" containing the MCMC samples.
}
\description{
Run the Gibbs sampler using a pre-defined mrdltm_model object.
}
