% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrdltm_model.R
\name{mrdltm_model}
\alias{mrdltm_model}
\title{Define MR-DLTM Specification}
\usage{
mrdltm_model(observations, n_topic = 3, Gt = NULL, ...)
}
\arguments{
\item{observations}{A list containing:
\itemize{
\item \code{data}: data.frame with (cust, item, time, y_cit)
\item \code{x_it}: 3D array of marketing covariates (item, time, n_var)
\item \code{Dc}: matrix of customer covariates (n_cust, p_dim). If NULL, defaults to local level.
}}

\item{n_topic}{Number of latent topics (Z)}

\item{Gt}{System matrix (p_dim x p_dim). Defaults to a local level model.}

\item{...}{Additional hyperparameters for priors.}
}
\value{
An object of class "mrdltm_model"
}
\description{
Constructor for an MR-DLTM object.
}
