% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrdltm_mcmc.R
\name{mrdltm_mcmc}
\alias{mrdltm_mcmc}
\title{Gibbs Sampler for MR-DLTM}
\usage{
mrdltm_mcmc(
  model,
  iter = 2000,
  burnin = 1000,
  thin = 1,
  prob_sample_ratio = 0.1,
  quiet = TRUE
)
}
\arguments{
\item{model}{An object of class "mrdltm_model".}

\item{iter}{Total MCMC iterations.}

\item{burnin}{Burn-in iterations.}

\item{thin}{MCMC thin setting.}

\item{prob_sample_ratio}{Numeric between 0 and 1. Proportion of observations to sample for
storing allocation probabilities. Lower values reduce memory usage. Default is 0.1 (10\%).}

\item{quiet}{A flag to show a progress-bar}
}
\value{
A list of class "mrdltm_mcmc" containing the MCMC samples.
}
\description{
Run the Gibbs sampler using a pre-defined mrdltm_model object.
Allocation probabilities are stored for post-hoc label switching correction
using Stephens' algorithm.
}
