% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{reorder_mrdltm}
\alias{reorder_mrdltm}
\title{Post-hoc Label Switching Correction for MR-DLTM}
\usage{
reorder_mrdltm(res, burnin = 0)
}
\arguments{
\item{res}{A list containing MCMC samples from mrdltm_mcmc.}

\item{burnin}{Integer. Number of ORIGINAL iterations to discard.}
}
\value{
A list of MCMC samples with corrected labels.
}
\description{
Aligns topic labels across MCMC iterations using \code{label.switching} package.
Automatically selects the algorithm based on the availability of 'z_cit':
\itemize{
\item If \code{store_z = TRUE}: Uses the \strong{ECR algorithm}. Efficient for large Z but requires memory for Z.
\item If \code{store_z = FALSE}: Uses the \strong{PRA algorithm}. Standard method using parameters, slower for large Z.
}
}
