% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{reorder_mrdltm}
\alias{reorder_mrdltm}
\title{Post-hoc Label Switching Correction for MR-DLTM}
\usage{
reorder_mrdltm(res, burnin = 0)
}
\arguments{
\item{res}{A list containing MCMC samples from mrdltm_mcmc.}

\item{burnin}{Integer. Number of initial iterations to discard.}
}
\value{
A list of MCMC samples with corrected labels and adjusted baselines.
}
\description{
Aligns topic labels across MCMC iterations using the Pivot Relabelling Algorithm (PRA).
This function also ensures the structural baseline constraint (Topic Z = 0)
is preserved by dynamically shifting the relative parameters (alpha and eta).
}
